+++ src/main/php/lang/ast/emit/PHP72.class.php
@@ -1,5 +1,7 @@
 <?php namespace lang\ast\emit;
 
+use lang\ast\types\{IsUnion, IsFunction, IsArray, IsMap, IsNullable, IsValue, IsLiteral};
+
 /**
  * PHP 7.2 syntax
  *
@@ -9,7 +11,19 @@ class PHP72 extends PHP {
   use OmitPropertyTypes;
   use RewriteNullCoalesceAssignment, RewriteLambdaExpressions, RewriteClassOnObjects;
 
-  protected $unsupported= [
-    'mixed'    => null,
-  ];
+  /** Sets up type => literal mappings */
+  public function __construct() {
+    $this->literals= [
+      IsArray::class    => function($t) { return 'array'; },
+      IsMap::class      => function($t) { return 'array'; },
+      IsFunction::class => function($t) { return 'callable'; },
+      IsValue::class    => function($t) { return $t->literal(); },
+      IsNullable::class => function($t) { $l= $this->literal($t->element); return null === $l ? null : '?'.$l; },
+      IsUnion::class    => function($t) { return null; },
+      IsLiteral::class  => function($t) {
+        $l= $t->literal();
+        return 'mixed' === $l ? null : $l;
+      },
+    ];
+  }
 }